# chronivaro-rest/src/main/java/ch/eitchnet/chronivaro/rest/resource/HolidayCalendarsResource.java
package ch.eitchnet.chronivaro.rest.resource;

import ch.eitchnet.chronivaro.core.service.CreateHolidayCalendarService;
import ch.eitchnet.chronivaro.core.service.CreateHolidayService;
import ch.eitchnet.chronivaro.core.service.RemoveHolidayCalendarService;
import ch.eitchnet.chronivaro.core.service.RemoveHolidayService;
import ch.eitchnet.chronivaro.rest.dto.ChronivaroMapper;
import ch.eitchnet.chronivaro.rest.dto.HolidayCalendarDto;
import ch.eitchnet.chronivaro.rest.dto.HolidayDto;
import com.google.gson.JsonObject;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.ws.rs.*;
import jakarta.ws.rs.core.Context;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import li.strolch.model.Resource;
import li.strolch.model.Tags;
import li.strolch.persistence.api.StrolchTransaction;
import li.strolch.privilege.model.Certificate;
import li.strolch.service.StringArgument;
import li.strolch.service.api.ServiceHandler;
import li.strolch.service.api.ServiceResult;

import java.util.List;

import static ch.eitchnet.chronivaro.core.model.ChronivaroConstants.*;
import static li.strolch.rest.StrolchRestfulConstants.STROLCH_CERTIFICATE;

@Path("chronivaro/v1/admin/holiday-calendars")
public class HolidayCalendarsResource {

	@GET
	@Produces(MediaType.APPLICATION_JSON)
	public Response getHolidayCalendars(@Context HttpServletRequest request,
			@QueryParam("offset") Integer offset, @QueryParam("limit") Integer limit) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		try (StrolchTransaction tx = ChronivaroRestHelper.openTx(cert)) {
			List<Resource> calendars = tx.streamResources(TYPE_HOLIDAY_CALENDAR).toList();
			return PaginationHelper.toPagedOrListResponse(calendars, offset, limit, ChronivaroMapper::holidayCalendarToDto);
		}
	}

	@GET
	@Path("{id}")
	@Produces(MediaType.APPLICATION_JSON)
	public Response getHolidayCalendar(@Context HttpServletRequest request, @PathParam("id") String id) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		try (StrolchTransaction tx = ChronivaroRestHelper.openTx(cert)) {
			Resource calendar = tx.getResourceBy(TYPE_HOLIDAY_CALENDAR, id, true);
			HolidayCalendarDto dto = ChronivaroMapper.holidayCalendarToDto(calendar);
			return ConcurrencyHelper.toResponseWithETag(calendar, dto);
		}
	}

	@GET
	@Path("{id}/holidays")
	@Produces(MediaType.APPLICATION_JSON)
	public Response getHolidays(@Context HttpServletRequest request, @PathParam("id") String id,
			@QueryParam("offset") Integer offset, @QueryParam("limit") Integer limit) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		try (StrolchTransaction tx = ChronivaroRestHelper.openTx(cert)) {
			List<Resource> holidays = tx
					.streamResources(TYPE_HOLIDAY)
					.filter(h -> h.getRelationId(PARAM_HOLIDAY_CALENDAR).equals(id))
					.toList();
			return PaginationHelper.toPagedOrListResponse(holidays, offset, limit, ChronivaroMapper::holidayToDto);
		}
	}

	@GET
	@Path("{id}/holidays/{holidayId}")
	@Produces(MediaType.APPLICATION_JSON)
	public Response getHoliday(@Context HttpServletRequest request, @PathParam("id") String id,
			@PathParam("holidayId") String holidayId) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		try (StrolchTransaction tx = ChronivaroRestHelper.openTx(cert)) {
			Resource holiday = tx.getResourceBy(TYPE_HOLIDAY, holidayId, true);
			return ConcurrencyHelper.toResponseWithETag(holiday, ChronivaroMapper.holidayToDto(holiday));
		}
	}

	@POST
	@Consumes(MediaType.APPLICATION_JSON)
	@Produces(MediaType.APPLICATION_JSON)
	public Response createHolidayCalendar(@Context HttpServletRequest request, String data) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		ServiceHandler serviceHandler = ChronivaroRestHelper.getServiceHandler();
		CreateHolidayCalendarService.HolidayCalendarArgument arg = ChronivaroRestHelper
				.createGson()
				.fromJson(data, CreateHolidayCalendarService.HolidayCalendarArgument.class);
		CreateHolidayCalendarService service = new CreateHolidayCalendarService();
		ServiceResult result = serviceHandler.doService(cert, service, arg);
		if (result.isOk()) {
			try (StrolchTransaction tx = ChronivaroRestHelper.openTx(cert)) {
				Resource calendar = tx
						.streamResources(TYPE_HOLIDAY_CALENDAR)
						.filter(r -> r
								.getString(ch.eitchnet.chronivaro.core.model.ChronivaroConstants.PARAM_NAME)
								.equals(arg.name))
						.findFirst()
						.orElse(null);
				if (calendar != null) {
					JsonObject json = new JsonObject();
					json.addProperty(Tags.Json.VALUE, calendar.getId());
					return ConcurrencyHelper.toResponseWithETag(calendar, json);
				}
			}
		}
		return ChronivaroRestHelper.toResponse(result);
	}

	@POST
	@Path("{id}/holidays")
	@Consumes(MediaType.APPLICATION_JSON)
	@Produces(MediaType.APPLICATION_JSON)
	public Response createHoliday(@Context HttpServletRequest request, @PathParam("id") String calendarId,
			String data) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		ServiceHandler serviceHandler = ChronivaroRestHelper.getServiceHandler();
		CreateHolidayService.HolidayArgument arg = ChronivaroRestHelper
				.createGson()
				.fromJson(data, CreateHolidayService.HolidayArgument.class);
		arg.holidayCalendarId = calendarId;
		ServiceResult result = serviceHandler.doService(cert, new CreateHolidayService(), arg);
		return ChronivaroRestHelper.toResponse(result);
	}

	@DELETE
	@Path("{id}")
	@Produces(MediaType.APPLICATION_JSON)
	public Response removeHolidayCalendar(@Context HttpServletRequest request, @PathParam("id") String id) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		try (StrolchTransaction tx = ChronivaroRestHelper.openTx(cert)) {
			Resource calendar = tx.getResourceBy(TYPE_HOLIDAY_CALENDAR, id, true);
			ConcurrencyHelper.validateIfMatch(request, calendar);
		}

		ServiceHandler serviceHandler = ChronivaroRestHelper.getServiceHandler();
		ServiceResult result = serviceHandler.doService(cert, new RemoveHolidayCalendarService(),
				new StringArgument(id));
		return ChronivaroRestHelper.toResponse(result);
	}

	@DELETE
	@Path("{calendarId}/holidays/{holidayId}")
	@Produces(MediaType.APPLICATION_JSON)
	public Response removeHoliday(@Context HttpServletRequest request, @PathParam("calendarId") String calendarId,
			@PathParam("holidayId") String holidayId) {
		Certificate cert = (Certificate) request.getAttribute(STROLCH_CERTIFICATE);
		try (StrolchTransaction tx = ChronivaroRestHelper.openTx(cert)) {
			Resource holiday = tx.getResourceBy(TYPE_HOLIDAY, holidayId, true);
			ConcurrencyHelper.validateIfMatch(request, holiday);
		}

		ServiceHandler serviceHandler = ChronivaroRestHelper.getServiceHandler();
		ServiceResult result = serviceHandler.doService(cert, new RemoveHolidayService(),
				new StringArgument(holidayId));
		return ChronivaroRestHelper.toResponse(result);
	}
}