# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chronivaro REST API
  description: |
    REST API specification for Chronivaro, a Resource-Order-Activity time tracking and absence management system built on Strolch.
    Provides endpoints for time tracking, timers, presence overview, absence requests, period closing, reports, and administrative management.
  version: 0.1.0-SNAPSHOT
  contact:
    name: Chronivaro
    url: https://github.com/eitch/Chronivaro
servers:
  - url: /
    description: Current server instance
paths:
...
  /chronivaro/v1/work-entries/day/{date}:
    get:
      summary: Get work entries for day
      description: Retrieves all recorded work entries for a specific date (YYYY-MM-DD).
      operationId: getWorkEntriesForDay
      tags:
        - Time Tracking
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          description: Target date formatted as YYYY-MM-DD
          schema:
            type: string
            format: date
            example: '2026-08-18'
        - name: employeeId
          in: query
          description: Employee ID (requires supervisor/admin privilege if querying another employee)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Work entries retrieved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkEntryDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /chronivaro/v1/summary/day/{date}:
    get:
      summary: Get day summary
      description: Returns target minutes, actual recorded minutes, holiday credits, absences, computed balance, and gap-based breaks.
      operationId: getDaySummary
      tags:
        - Summaries
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          description: Target date formatted as YYYY-MM-DD
          schema:
            type: string
            format: date
            example: '2026-08-18'
        - name: employeeId
          in: query
          description: Employee ID (optional, defaults to current user's employee)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Day summary retrieved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DaySummaryDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /chronivaro/v1/admin/holiday-calendars:
    get:
      summary: List holiday calendars
      description: Returns configured public holiday calendars.
      operationId: getHolidayCalendars
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Holiday calendars retrieved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/HolidayCalendarDto'
                  - $ref: '#/components/schemas/PagedHolidayCalendarDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create holiday calendar
      description: Creates a new regional holiday calendar.
      operationId: createHolidayCalendar
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayCalendarCreateDto'
      responses:
        '200':
          description: Holiday calendar created successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayCalendarDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/admin/holiday-calendars/{id}:
    get:
      summary: Get holiday calendar by ID
      description: Returns details of a specific holiday calendar.
      operationId: getHolidayCalendarById
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
      responses:
        '200':
          description: Holiday calendar retrieved successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayCalendarDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update holiday calendar
      description: Updates holiday calendar name. Supports optimistic concurrency via `If-Match`.
      operationId: updateHolidayCalendar
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayCalendarCreateDto'
      responses:
        '200':
          description: Holiday calendar updated successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayCalendarDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Remove holiday calendar
      description: Removes a holiday calendar. Supports optimistic concurrency via `If-Match`.
      operationId: removeHolidayCalendar
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      responses:
        '200':
          description: Holiday calendar removed successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/admin/holiday-calendars/{id}/holidays:
    get:
      summary: List holidays in calendar
      description: Returns individual holiday definitions for the calendar with optional year filtering.
      operationId: getHolidays
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
        - name: year
          in: query
          required: false
          description: Year filter (e.g., 2026)
          schema:
            type: integer
            example: 2026
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Holidays list retrieved successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/HolidayDto'
                  - $ref: '#/components/schemas/PagedHolidayDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Add holiday to calendar
      description: Defines a holiday date, name, and credit factor (e.g., 1.0 for full day, 0.5 for half day).
      operationId: createHoliday
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayCreateDto'
      responses:
        '200':
          description: Holiday added successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /chronivaro/v1/admin/holiday-calendars/{id}/holidays/{holidayId}:
    get:
      summary: Get holiday by ID
      description: Returns details of a specific holiday definition.
      operationId: getHolidayById
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
        - name: holidayId
          in: path
          required: true
          description: Holiday identifier
          schema:
            type: string
      responses:
        '200':
          description: Holiday retrieved successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update holiday
      description: Updates holiday date, name, or credit factor. Supports optimistic concurrency via `If-Match`.
      operationId: updateHoliday
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
        - name: holidayId
          in: path
          required: true
          description: Holiday identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayCreateDto'
      responses:
        '200':
          description: Holiday updated successfully
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidayDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Remove holiday
      description: Removes a holiday from calendar. Supports optimistic concurrency via `If-Match`.
      operationId: removeHoliday
      tags:
        - Admin - Holiday Calendars
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Holiday calendar identifier
          schema:
            type: string
        - name: holidayId
          in: path
          required: true
          description: Holiday identifier
          schema:
            type: string
        - $ref: '#/components/parameters/IfMatchParam'
      responses:
        '200':
          description: Holiday removed successfully
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/X-Correlation-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # Admin - Schedule Templates
  # ==========================================
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: strolch.authorization
      description: Strolch session cookie containing the user authentication token.
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token containing the Strolch user certificate ID.
  headers:
    ETag:
      description: Entity version validator for optimistic concurrency control (e.g., W/"1" or "1").
      schema:
        type: string
        example: '"1"'
...
    X-Correlation-Id:
      description: Unique identifier for tracking and correlating requests across logs and client telemetry.
      schema:
        type: string
        format: uuid
        example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Zero-based item offset for pagination (defaults to 0).
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
        example: 0
    LimitParam:
      name: limit
      in: query
      description: Number of items to return (defaults to 50, maximum 1000).
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
        example: 50
    IfMatchParam:
      name: If-Match
      in: header
      required: false
      description: Target entity version for optimistic concurrency control.
      schema:
        type: string
        example: '"1"'
  responses:
    BadRequest:
      description: Bad request - Validation failure or malformed payload
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Unauthorized:
      description: Unauthorized - Authentication is missing or invalid
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Forbidden:
      description: Forbidden - Insufficient user privileges or role permissions
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    NotFound:
      description: Not found - Requested resource does not exist
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    Conflict:
      description: Conflict - Concurrency mismatch (stale If-Match) or business state violation
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
    InternalServerError:
      description: Internal server error - Unexpected system failure
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'
  schemas:
...
    ErrorDto:
      type: object
      required:
        - error
        - code
        - message
        - correlationId
      properties:
        error:
          type: boolean
          example: true
        code:
          type: string
          example: 'VALIDATION_ERROR'
        message:
          type: string
          example: 'Invalid input parameters'
        correlationId:
          type: string
          example: '4a1b0289-53e7-4b72-bdf4-f8b898ecb817'
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    FieldErrorDto:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: 'email'
        message:
          type: string
          example: 'Email address is invalid'
    SuccessResponseDto:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Operation completed successfully'
...
    WorkEntryDto:
      type: object
      properties:
        id:
          type: string
          example: 'we-20260818-01'
        employeeId:
          type: string
          example: 'emp-001'
        start:
          type: string
          format: date-time
          example: '2026-08-18T08:00:00+02:00'
        end:
          type: string
          format: date-time
          example: '2026-08-18T12:00:00+02:00'
        durationMinutes:
          type: integer
          example: 240
        source:
          type: string
          enum: [TIMER, MANUAL, API]
          example: 'TIMER'
        comment:
          type: string
          nullable: true
          example: 'Morning development session'
        createdBy:
          type: string
          example: 'admin'
        workingLocation:
          type: string
          enum: [OFFICE, REMOTE, CLIENT, TRANSIT]
          example: 'OFFICE'
        modified:
          type: boolean
          example: false
...
    WorkEntryRangeDto:
      type: object
      properties:
        id:
          type: string
          example: 'we-20260818-01'
        start:
          type: string
          example: '08:00'
        end:
          type: string
          example: '12:00'
        durationMinutes:
          type: integer
          example: 240
        source:
          type: string
          enum: [TIMER, MANUAL, API]
          example: 'TIMER'
        createdBy:
          type: string
          example: 'admin'
        modified:
          type: boolean
          example: false
    BreakRangeDto:
      type: object
      properties:
        start:
          type: string
          format: date-time
          example: '2026-08-18T12:00:00+02:00'
        end:
          type: string
          format: date-time
          example: '2026-08-18T12:45:00+02:00'
        durationMinutes:
          type: integer
          example: 45

    # ------------------------------------------
    # Summaries
    # ------------------------------------------
    DaySummaryDto:
      type: object
      properties:
        date:
          type: string
          format: date
          example: '2026-08-18'
        state:
          type: string
          enum: [REGULAR, ABSENT, HOLIDAY, WEEKEND, OFF]
          example: 'REGULAR'
        stateLabel:
          type: string
          example: 'Regular Workday'
        targetMinutes:
          type: integer
          example: 480
        actualMinutes:
          type: integer
          example: 510
        holidayMinutes:
          type: integer
          example: 0
        absenceMinutes:
          type: integer
          example: 0
        isOff:
          type: boolean
          example: false
        balance:
          type: integer
          example: 30
        workingLocation:
          type: string
          enum: [OFFICE, REMOTE, CLIENT, TRANSIT]
          example: 'OFFICE'
        workEntries:
          type: array
          items:
            $ref: '#/components/schemas/WorkEntryRangeDto'
        breaks:
          type: array
          items:
            $ref: '#/components/schemas/BreakRangeDto'
...
    HolidayCalendarDto:
      type: object
      properties:
        id:
          type: string
          example: 'hol-ch-zh'
        name:
          type: string
          example: 'Canton Zurich'
    HolidayCalendarCreateDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: 'Canton Zurich'
    PagedHolidayCalendarDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HolidayCalendarDto'
        offset:
          type: integer
          example: 0
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 2
        size:
          type: integer
          example: 2
    HolidayDto:
      type: object
      properties:
        id:
          type: string
          example: 'hol-2026-08-01'
        holidayCalendarId:
          type: string
          example: 'hol-ch-zh'
        date:
          type: string
          format: date
          example: '2026-08-01'
        name:
          type: string
          example: 'Swiss National Day'
        creditFactor:
          type: number
          format: double
          example: 1.0
    HolidayCreateDto:
      type: object
      required:
        - date
        - name
        - creditFactor
      properties:
        date:
          type: string
          format: date
          example: '2026-08-01'
        name:
          type: string
          example: 'Swiss National Day'
        creditFactor:
          type: number
          format: double
          example: 1.0
    PagedHolidayDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HolidayDto'
        offset:
          type: integer
          example: 0
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 12
        size:
          type: integer
          example: 12
...